requirements.txt
Cython==0.21.2
celery[redis]==3.1.17
elasticsearch==1.3.0
click==2.1
coverage==4.0a1
Flask==0.10.1
Flask-Testing==0.4.2
iso8601==0.1.10
lxml==3.3.5
requests==2.5.1
mock==1.0.1
msgpack-python==0.4.2
ndg-httpsclient==0.3.2
nose==1.3.4
pyOpenSSL==0.14
pyasn1==0.1.7
Sphinx==1.2.2
sphinxcontrib-httpdomain==1.2.1
sphinx-rtd-theme==0.1.6
suds==0.4
translitcodec==0.4.0
BeautifulSoup==3.2.1
Pillow==2.7.0
-e git+https://github.com/mikeboers/PyAV.git@40b3cb85f1d1b1823fdb24dd0aaa01c14d005d3c#egg=av
spyne==2.11.0
